LayoutTests/http/tests/contentextensions/block-object-element.html

<script>testRunner?.dumpAsText();</script>
<body>
<object data="http://127.0.0.1:8000/contentextensions/resources/should-load.html?should-be-blocked" type="text/html" width="100" height="100"></object>
<object data="http://127.0.0.1:8000/media/resources/test.pdf?should-be-blocked" type="application/pdf" width="100" height="100"></object>
</body>
