LayoutTests/http/tests/contentextensions/block-object-element-expected.txt

CONSOLE MESSAGE: Content blocker prevented frame displaying http://127.0.0.1:8000/contentextensions/block-object-element.html from loading a resource from http://127.0.0.1:8000/contentextensions/resources/should-load.html?should-be-blocked
CONSOLE MESSAGE: Content blocker prevented frame displaying http://127.0.0.1:8000/contentextensions/block-object-element.html from loading a resource from http://127.0.0.1:8000/media/resources/test.pdf?should-be-blocked
